requirements.txt
Django==2.2
PyJWT==1.7.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
django-ajax-selects==1.9.1
django-js-asset==1.1.0
django-mptt==0.11.0
django-suit==0.2.28
docutils==0.14
idna==2.8
mysqlclient==1.4.6
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
pre-commit
